@@ -1,28 +1,27 @@
 {
-  "name": "react-portfolio-website",
+  "name": "personal-portfolio",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emailjs/browser": "^3.4.0",
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-brands-svg-icons": "^6.0.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "animate.css": "^4.1.1",
-    "firebase": "^9.8.3",
-    "gsap-trial": "^3.9.0",
-    "leaflet": "^1.7.1",
-    "loaders.css": "^0.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-leaflet": "^3.2.5",
-    "react-loaders": "^3.0.1",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "sass": "^1.49.9",
+    "bootstrap": "^5.1.3",
+    "cors": "^2.8.5",
+    "express": "^4.18.1",
+    "nodemailer": "^6.7.5",
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.4.0",
+    "react-bootstrap-icons": "^1.8.2",
+    "react-dom": "^18.1.0",
+    "react-mailchimp-subscribe": "^2.1.3",
+    "react-multi-carousel": "^2.8.1",
+    "react-on-screen": "^2.1.1",
+    "react-responsive-carousel": "^3.2.23",
+    "react-router-dom": "^6.3.0",
+    "react-router-hash-link": "^2.4.3",
+    "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
